Fix issue#1

+++ src/helpers/vscode.ts
@@ -11,27 +11,43 @@ export const getConfigProperty = <T extends string>(property: Properties): T =>
 export const createFile = async (args: CreateFile): Promise<string | null> => {
   try {
     const file = path.join(args.path, fixExtension(args.name, args.extension));
-    vscode.window.showInformationMessage(`Create file ${file}`);
+    if (fs.existsSync(file)) {
+      vscode.window.showErrorMessage(
+        `File ${path.basename(file)} already exist. Full path: ${file}`
+      );
+      return null;
+    }
     fs.writeFileSync(file, "");
+    vscode.window.showInformationMessage(`Create file ${file}`);
     return file;
   } catch (error) {
     return null;
   }
 };
 
+const extractBasename = async (file: string): Promise<string> => {
+  const stat = await fs.promises.lstat(file);
+  if (stat.isFile()) {
+    return path.dirname(file);
+  }
+  return file;
+};
+
 export const getContextPath = async (
   ctx: vscode.ExtensionContext,
   args: ArgsPrompt
 ): Promise<string> => {
+  const absolutePath = ctx.asAbsolutePath(".");
   if (args) {
-    let contextRootPath = args.fsPath || ""; // ||
+    let contextRootPath = await extractBasename(args.fsPath || "");
+    console.log("PATH", contextRootPath);
     if (contextRootPath === "") {
       const tmpdir =
         (await vscode.window.showInputBox({
           prompt: "Directory",
           ignoreFocusOut: true,
         })) || "";
-      return path.join(ctx.asAbsolutePath("."), tmpdir);
+      return path.join(absolutePath, tmpdir);
     }
     return contextRootPath;
   }
@@ -40,7 +56,7 @@ export const getContextPath = async (
       prompt: "Directory",
       ignoreFocusOut: true,
     })) || ".";
-  return path.join(ctx.asAbsolutePath("."), tmpdir);
+  return path.join(absolutePath, await extractBasename(tmpdir));
 };
 
 export const getFromPrompt = async ({
@@ -56,20 +72,22 @@ export const openFile = async (file: string): Promise<void> => {
 };
 
 export const writeTemplate = (filename: string, code: string) => {
-  vscode.workspace.openTextDocument(vscode.Uri.parse(filename)).then(
-    (textDocument: vscode.TextDocument) => {
-      vscode.window
-        .showTextDocument(textDocument, 1, false)
-        .then((textEditor) => {
-          textEditor.edit((edit) => {
-            edit.insert(new vscode.Position(0, 0), code);
+  if (code !== "") {
+    vscode.workspace.openTextDocument(vscode.Uri.parse(filename)).then(
+      (textDocument: vscode.TextDocument) => {
+        vscode.window
+          .showTextDocument(textDocument, 1, false)
+          .then((textEditor) => {
+            textEditor.edit((edit) => {
+              edit.insert(new vscode.Position(0, 0), code);
+            });
           });
-        });
-    },
-    (error: any) => {
-      console.error(error);
-    }
-  );
+      },
+      (error: any) => {
+        console.error(error);
+      }
+    );
+  }
 };
 
 export const formatDocument = () =>
